* util.c (ruby_add_suffix): suppress a warning.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@28806 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ util.c
@@ -277,7 +277,7 @@ ruby_add_suffix(VALUE str, const char *suffix)
 
     name = StringValueCStr(str);
     slen = strlen(name);
-    if (slen > sizeof(buf) - 1)
+    if (slen > (long)(sizeof(buf) - 1))
 	rb_fatal("Cannot do inplace edit on long filename (%ld characters)",
 		 slen);
